Llama Dash
While exploring the LCD capabilities, I created a "Space Invader" type game. The player gets to control a llama in a 2D plane track while a ufo flies above them trying to shoot lazer beams to kill the llama. Two values are displayed on the top edges: on the left, we have the survival timer. This keeps track of how long the player has survived. On the right, we have the number of lazer beams that have been shot down by the ufo and successfully dodged. The longer you live, the better you score. Once the game is over, you must press the reset button to start a new game. Have fun!
PS- There is one glitch I still haven't been able to resolve, feel free to try a crack at it. When the player crosses the ufo for the first time, the counter tick goes up by one on the top right (ufo must be away from the 3 boxes from the right to see), but the ship does not fire a beam. It works perfectly after that.